Implement signin/signup flow with main page

+++ app/src/main/java/com/ubidel/ubicash/api/model/UserDetails.java
@@ -30,7 +30,7 @@ public class UserDetails {
     @Setter
     @SerializedName("phone")
     @Expose
-    public Integer phone;
+    public Double phone;
     @Getter
     @Setter
     @SerializedName("verified")
